enhancement: fix tweet styling
